refactor: move helpers to dedicated file
